HHH-14826 Extract issue specific tests from OneToOneCacheTest
